Branch refactoring and some changes inn the code to refactor the branch
