+++ Cluedo/src/main/java/cluedo/logic/players/Color.java
