+++ Cluedo/src/main/java/cluedo/logic/players/Player.java
@@ -1,0 +1,14 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package cluedo.logic.players;
+
+/**
+ *
+ * @author Bence
+ */
+public class Player {
+    
+}
